added folder gui for non swing gui things. split DbgPacker gui / util
