+++ src/main/java/net/pms/dlna/LibraryScanner.java
@@ -18,7 +18,6 @@
 package net.pms.dlna;
 
 import net.pms.PMS;
-import net.pms.newgui.SharedContentTab;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -47,7 +46,7 @@ public class LibraryScanner {
 			scanner = new Thread(scan, "Library Scanner");
 			scanner.setPriority(Thread.MIN_PRIORITY);
 			scanner.start();
-			SharedContentTab.setScanLibraryBusy();
+			PMS.get().getFrame().setScanLibraryStatus(true, true);
 		}
 	}
