+++ src/main/java/net/pms/gui/EConnectionState.java
@@ -1,7 +1,7 @@
 /*
  * This file is part of Universal Media Server, based on PS3 Media Server.
  *
- * This program is free software; you can redistribute it and/or
+ * This program is a free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
  * as published by the Free Software Foundation; version 2
  * of the License only.
@@ -15,14 +15,8 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package net.pms.newgui;
+package net.pms.gui;
 
-import net.pms.configuration.RendererConfiguration;
-import net.pms.util.BasicPlayer;
-
-public interface IRendererGuiListener {
-	public void updateRenderer(final RendererConfiguration renderer);
-	public void setActive(final boolean active);
-	public void delete();
-	public void refreshPlayerState(BasicPlayer.State state);
+public enum EConnectionState {
+	SEARCHING, CONNECTED, DISCONNECTED, BLOCKED, UNKNOWN
 }
