+++ src/main/java/net/pms/network/webinterfaceserver/handlers/ConsoleHandler.java
@@ -27,9 +27,9 @@ import java.util.Set;
 import net.pms.PMS;
 import net.pms.configuration.PmsConfiguration;
 import net.pms.dlna.LibraryScanner;
-import net.pms.newgui.DbgPacker;
 import net.pms.network.webinterfaceserver.WebInterfaceServerUtil;
 import net.pms.network.webinterfaceserver.WebInterfaceServerHttpServer;
+import net.pms.util.DbgPacker;
 import net.pms.util.PropertiesUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
