+++ src/main/java/net/pms/newgui/DbgPacker.java
@@ -11,10 +11,7 @@ import java.io.IOException;
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
 import java.util.Date;
-import java.util.HashMap;
-import java.util.LinkedHashMap;
 import java.util.Map;
-import java.util.Set;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipOutputStream;
 import javax.swing.*;
@@ -22,36 +19,12 @@ import javax.swing.filechooser.FileFilter;
 import javax.swing.plaf.metal.MetalIconFactory;
 import net.pms.Messages;
 import net.pms.PMS;
-import net.pms.configuration.DeviceConfiguration;
-import net.pms.configuration.PmsConfiguration;
-import net.pms.configuration.RendererConfiguration;
-import net.pms.logging.LoggingConfig;
 import net.pms.newgui.components.CustomJButton;
 import net.pms.util.FileUtil;
-import org.apache.commons.lang3.StringUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-public class DbgPacker implements ActionListener {
-	private static final Logger LOGGER = LoggerFactory.getLogger(DbgPacker.class);
-
-	private final LinkedHashMap<File, JCheckBox> items;
-	private String defaultLogFile, zippedLogFile;
+public class DbgPacker extends net.pms.util.DbgPacker implements ActionListener {
+	private String zippedLogFile;
 	private CustomJButton openZip;
-
-	public DbgPacker() {
-		items = new LinkedHashMap<>();
-
-		HashMap<String, String> logFilePaths = LoggingConfig.getLogFilePaths();
-		if (!logFilePaths.isEmpty()) {
-			defaultLogFile = LoggingConfig.getLogFilePaths().get("default.log");
-			if (defaultLogFile == null) {
-				// Just get the path of one of the files as we can't find the default
-				Map.Entry<String, String> entry = logFilePaths.entrySet().iterator().next();
-				defaultLogFile = entry.getValue();
-			}
-		}
-	}
 
 	public JComponent config() {
 		poll();
@@ -62,13 +35,15 @@ public class DbgPacker implements ActionListener {
 		c.ipadx = 5;
 		c.gridx = 0;
 		c.gridy = 0;
-		for (Map.Entry<File, JCheckBox> item : items.entrySet()) {
+		for (Map.Entry<File, Object> item : items.entrySet()) {
 			File file = item.getKey();
 			boolean exists = file.exists();
-			JCheckBox box = item.getValue();
-			if (box == null) {
+			JCheckBox box;
+			if (item.getValue() == null) {
 				box = new JCheckBox(file.getName(), exists);
 				item.setValue(box);
+			} else {
+				box = (JCheckBox) item.getValue();
 			}
 			if (!exists) {
 				box.setSelected(false);
@@ -103,60 +78,6 @@ public class DbgPacker implements ActionListener {
 		return top;
 	}
 
-	private void poll() {
-		// call the client callbacks
-		PmsConfiguration configuration = PMS.getConfiguration();
-
-		// check dbgpack property in UMS.conf
-		LOGGER.debug("Checking dbgpack property in UMS.conf");
-		String f = (String) configuration.getCustomProperty("dbgpack");
-		if (f != null) {
-			add(f.split(","));
-		}
-
-		// add confs of connected renderers
-		for (RendererConfiguration r : RendererConfiguration.getConnectedRenderersConfigurations()) {
-			add(r.getFile());
-			if (((DeviceConfiguration) r).isCustomized()) {
-				add(((DeviceConfiguration) r).getParentFile());
-			}
-		}
-
-		// add core items with the default logfile last (LinkedHashMap preserves
-		// insertion order)
-		String profileDirectory = configuration.getProfileDirectory();
-
-		// add virtual folders file if it exists
-		String vfolders = configuration.getVirtualFoldersFile();
-		if (StringUtils.isNotEmpty(vfolders)) {
-			add(new File(profileDirectory, vfolders));
-		}
-
-		add(new File(profileDirectory, "WEB.conf"));
-		add(new File(configuration.getProfilePath()));
-		if (defaultLogFile != null && !defaultLogFile.isEmpty()) {
-			add(new File(defaultLogFile + ".prev.zip"));
-			add(new File(defaultLogFile + ".zip"));
-			add(new File(defaultLogFile));
-		}
-	}
-
-	private void add(String[] files) {
-		for (String file : files) {
-			add(new File(file));
-		}
-	}
-
-	private void add(File file) {
-		if (file != null) {
-			LOGGER.debug("adding {}", file.getAbsolutePath());
-			try {
-				items.put(file.getCanonicalFile(), null);
-			} catch (IOException e) {
-			}
-		}
-	}
-
 	private void writeToZip(ZipOutputStream out, File f) throws Exception {
 		byte[] buf = new byte[1024];
 		int len;
@@ -171,11 +92,6 @@ public class DbgPacker implements ActionListener {
 			}
 			out.closeEntry();
 		}
-	}
-
-	public Set<File> getItems() {
-		poll();
-		return items.keySet();
 	}
 
 	private boolean saveDialog() {
@@ -225,8 +141,8 @@ public class DbgPacker implements ActionListener {
 		}
 		try {
 			try (ZipOutputStream zos = new ZipOutputStream(new FileOutputStream(zippedLogFile))) {
-				for (Map.Entry<File, JCheckBox> item : items.entrySet()) {
-					if (item.getValue().isSelected()) {
+				for (Map.Entry<File, Object> item : items.entrySet()) {
+					if (item.getValue() != null && ((JCheckBox) item.getValue()).isSelected()) {
 						File file = item.getKey();
 						LOGGER.debug("Packing {}", file.getAbsolutePath());
 						writeToZip(zos, file);
