+++ src/main/java/net/pms/newgui/TracesTab.java
@@ -17,6 +17,7 @@
  */
 package net.pms.newgui;
 
+import net.pms.gui.ViewLevel;
 import ch.qos.logback.classic.Level;
 import com.jgoodies.forms.builder.PanelBuilder;
 import com.jgoodies.forms.layout.CellConstraints;
