+++ src/main/java/net/pms/util/APIUtils.java
@@ -68,7 +68,7 @@ import net.pms.dlna.DLNAMediaVideoMetadata;
 import net.pms.dlna.DLNAThumbnail;
 import net.pms.image.ImageFormat;
 import net.pms.image.ImagesUtil.ScaleType;
-import net.pms.newgui.IFrame;
+import net.pms.gui.IFrame;
 import net.pms.util.OpenSubtitle.OpenSubtitlesBackgroundWorkerThreadFactory;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
