+++ src/test/java/net/pms/newgui/ViewLevelTest.java
@@ -17,6 +17,7 @@
  */
 package net.pms.newgui;
 
+import net.pms.gui.ViewLevel;
 import ch.qos.logback.classic.Level;
 import ch.qos.logback.classic.Logger;
 import ch.qos.logback.classic.LoggerContext;
